refactoring graph files
